[cli/update] update e2e tests to use file based update mechanic
