+++ e2e/tests/api_update.yaml
@@ -1,0 +1,11 @@
+kind: API
+name: api-test-update
+authentication: public
+enabled: true
+function: func-nodejs6
+methods:
+  - GET
+protocols:
+  - https
+uris:
+  - /goodbye
