+++ e2e/tests/image_not_found_update.yaml
@@ -1,0 +1,9 @@
+# Test when a base image does not exist
+dockerUrl: vmware/dispatch-python3-base:0.0.1-dev1
+groups: 
+kind: BaseImage
+language: python3
+name: python3-bases
+tags:
+- key: update
+  value: test
