+++ e2e/tests/secret_update.yaml
@@ -1,0 +1,7 @@
+kind: Secret
+name: open-sesame
+secrets:
+  password: OpenSesameStreet
+tags:
+  - key: role
+    value: test
